figured out bug: description
Because the two players have the same names when launching the same exe twice. When the second player disconnects, the server searches through the list of users to find a specific username, but because they both have the same username, the server actually deletes both of the users from the network, which is why player 1 cannot move.